+++ src/flask_caching/backends/memcache.py
@@ -11,6 +11,7 @@ The memcache caching backend.
 
 import pickle
 import re
+from functools import partial
 from typing import Any
 
 from cachelib import MemcachedCache as CachelibMemcachedCache
@@ -25,8 +26,8 @@ class MemcachedCache(BaseCache, CachelibMemcachedCache):
     """A cache that uses memcached as backend.
 
     The first argument can either be an object that resembles the API of a
-    :class:`memcache.Client` or a tuple/list of server addresses. In the
-    event that a tuple/list is passed, Werkzeug tries to import the best
+    ``memcache.Client`` or a tuple/list of server addresses. In the
+    event that a tuple/list is passed, CacheLib tries to import the best
     available memcache library.
 
     This cache looks into the following packages/modules to find bindings for
@@ -37,23 +38,36 @@ class MemcachedCache(BaseCache, CachelibMemcachedCache):
         - ``memcached``
         - ``libmc``
 
-    Implementation notes:  This cache backend works around some limitations in
-    memcached to simplify the interface.  For example unicode keys are encoded
-    to utf-8 on the fly.  Methods such as :meth:`~BaseCache.get_dict` return
+    Implementation notes:
+    This cache backend works around some limitations in memcached to
+    simplify the interface. For example unicode keys are encoded to
+    UTF-8 on the fly. Methods such as :meth:`~.BaseCache.get_dict` return
     the keys in the same format as passed.  Furthermore all get methods
     silently ignore key errors to not cause problems when untrusted user data
     is passed to the get methods which is often the case in web applications.
+    This cache doesn't have a serializer since the underlying memcached client
+    libraries handle serialization internally."
 
     :param servers: a list or tuple of server addresses or alternatively
-                    a :class:`memcache.Client` or a compatible client.
+        a ``memcache.Client`` or a compatible client.
     :param default_timeout: the default timeout that is used if no timeout is
-                            specified on :meth:`~BaseCache.set`. A timeout of
-                            0 indicates that the cache never expires.
+        specified on :meth:`~BaseCache.set`. A timeout of
+        0 indicates that the cache never expires.
     :param key_prefix: a prefix that is added before all keys.  This makes it
-                       possible to use the same memcached server for different
-                       applications.  Keep in mind that
-                       :meth:`~BaseCache.clear` will also clear keys with a
-                       different prefix.
+        possible to use the same memcached server for different
+        applications.  Keep in mind that
+        :meth:`~.BaseCache.clear` will also clear keys with a
+        different prefix.
+    :param pool_size: the size of the connection pool.  This is only used if
+        the memcached client library supports connection pooling.
+
+        .. versionadded:: 2.5.0
+    :param pool_blocking: if the connection pool is exhausted, should the
+        client block until a connection is available or raise
+        an exception.  This is only used if the memcached
+        client library supports connection pooling.
+
+        .. versionadded:: 2.5.0
     """
 
     def __init__(
@@ -61,6 +75,8 @@ class MemcachedCache(BaseCache, CachelibMemcachedCache):
         servers: Any = None,
         default_timeout: int = 300,
         key_prefix: str | None = None,
+        pool_size: int = 1,
+        pool_blocking: bool = True,
     ) -> None:
         BaseCache.__init__(self, default_timeout=default_timeout)
         CachelibMemcachedCache.__init__(
@@ -68,6 +84,8 @@ class MemcachedCache(BaseCache, CachelibMemcachedCache):
             servers=servers,
             default_timeout=default_timeout,
             key_prefix=key_prefix,
+            pool_size=pool_size,
+            pool_blocking=pool_blocking,
         )
 
     @classmethod
@@ -91,6 +109,8 @@ class SASLMemcachedCache(MemcachedCache):
         key_prefix: str | None = None,
         username: str | None = None,
         password: str | None = None,
+        pool_size: int = 1,
+        pool_blocking: bool = True,
         **kwargs: Any,
     ) -> None:
         super().__init__(default_timeout=default_timeout)
@@ -100,9 +120,13 @@ class SASLMemcachedCache(MemcachedCache):
 
         import pylibmc  # # type: ignore pyright: ignore[reportMissingImports]
 
-        self._client = pylibmc.Client(
+        client = pylibmc.Client(
             servers, username=username, password=password, binary=True, **kwargs
         )
+        pool = pylibmc.ClientPool(client, pool_size)
+        reserve = partial(pool.reserve, block=pool_blocking)
+        self._client = pool
+        self._client_context = reserve
 
         self.key_prefix = key_prefix